refactor(server): begin refactoring server build to include express server creation
