+++ packages/core/config/webpack.config.js
@@ -67,6 +67,7 @@ module.exports = (env, argv) => {
       entry: server.getEntry(),
       output: server.getOutput(),
       module: {
+        noParse: [/nodeRequire/],
         rules: server.getRules(),
       },
       plugins: server.getPlugins(),
