+++ packages/core/config/webpack/plugins.js
@@ -1,7 +1,7 @@
 const webpack = require('webpack');
 const CleanPlugin = require('clean-webpack-plugin');
 const { StatsWriterPlugin } = require('webpack-stats-plugin');
-const { maybeResolveUserModule } = require('../../utils/userModule');
+const { maybeResolveUserModule } = require('../../utils/nodeRequire');
 const getEnv = require('../env');
 const { rootUrl } = require('../paths');
